春至/data-monitor/requirement.txt
requests==2.22.0
openpyxl==2.6.2
mysql-connector
ddt==1.2.1
rsa==4.6
APScheduler==3.6.3
etcd3==0.12.0
confluent-kafka==1.5.0
ws4py==0.5.1
websocket==0.2.1
websocket-client==0.57.0
pymongo==3.11.0